+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 3.0.5 (unreleased)
+# 3.0.5
 
 * Fix calling #select from a pipeline (#309).